FA-591: enforce trailing slash configuration
This commit configures `docusaurus` to use an explicit `trailingSlash`
configuration. Without this configuration, a file `doc.md` would be
rendered as `doc/index.html`, but docusaurus would accept links to it as
`/doc`. When such a link was served by `nginx`, it would automatically
(and correctly) perform a redirect to `/doc/`, as once rendered, the
user is requesting a directory. However, when such a link was processed
by the in-javascript loading, docusaurus would fail to resolve the file
and return an error.

This commit moves us to using the more strict `trailingSlash: true`
mode, meaning that links are always emitted in the form containing a
trailing slash. I did consider using `trailingSlash: false`, but this
would result in `doc.md` being rendered as `doc.html`, and we would
rather not have the file extension in URLs and avoiding the need for URL
rewrite helps keep the server config simple. Because of this move, a
number of the internal links within the documentation which linked to
URLs, rather than to specific documents, are now detected as invalid by
docusaurus during the build. This commit therefore also fixes those
links to refer to the origin documents, allowing docusuarus to resolve
and rewrite them successfully.