`/support`: Fix broken inter-doc links
On the support page, we use some HTML `a` tags styled as buttons, which
it seems docusaurus doesn't validate for broken links on build.
Consequently, they accidentally became broken. This commit fixes this in
a sticking-plaster fashion. We should consider migrating these buttons
to being of a form that can be statically checked by docusaurus.