getting-started: fix headings (#29)
An article should only have one top-level heading, as this is used to
generate the article title. In-article navigation requires sub-headings
to be second-level headings or below. With the in-article headings as
top-level headings, they do not appear in the in-article navigation and
cannot be linked to.

This commit modifies the in-article headings to be second-level headings
or below.