FA-643: ci: Unify deployment action (#12)
The deployment action should use the name of the current repository to
determine which version of the action to run. Otherwise, merging between
multiple remotes of the docs site (live, staging) is basically
impossible without having to constantly modify the deployment action or
having the deployment action running on the wrong instance.

This commit also pins the versions of the `appleboy` actions that we use.