Resolve time-travelling screenshots (#57)
* Restore old screenshots

Restore the previous version of the screenshot resources
which were changed in 3bdbf58 (#55), as we do not want the updated
screenshots to show for old versions of the manual.

The updated screenshots will be incorporated into the next version of
the manual in a subsequent commit.

* Add v1.5 screenshots

Re-add the screenshots that were reverted in the previous commit,
placing them in a separate subdirectory. Refer to the updated files
wherever they are used in the docs.

(cherry picked from commit 8ff596659035df1de8e561fa58bcf0a62219f0cd)